Fix them all
