+++ src/main/java/io/appium/java_client/proxy/Interceptor.java
@@ -24,119 +24,13 @@ import net.bytebuddy.implementation.bind.annotation.This;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.lang.ref.WeakReference;
 import java.lang.reflect.Method;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.Callable;
-import java.util.concurrent.Semaphore;
 
 public class Interceptor {
     private static final Logger logger = LoggerFactory.getLogger(Interceptor.class);
-    // Previously WeakHashMap has been used because of O(1) lookup performance, although
-    // we had to change it to a list, which has O(N). The reason for that is that
-    // maps implicitly call `hashCode` API on instances, which might not always
-    // work as expected for arbitrary proxies
-    private static final List<Pair<WeakReference<?>, Collection<MethodCallListener>>> LISTENERS = new LinkedList<>();
-    private static final Semaphore LISTENERS_GUARD = new Semaphore(1);
-
-    private static class Pair<K, V> {
-        private final K key;
-        private final V value;
-
-        public Pair(K key, V value) {
-            this.key = key;
-            this.value = value;
-        }
-
-        public K getKey() {
-            return key;
-        }
-
-        public V getValue() {
-            return value;
-        }
-    }
-
-
-    /**
-     * Assign listeners for the particular proxied instance.
-     *
-     * @param instance The proxied instance.
-     * @param listeners Collection of listeners.
-     * @return The same given instance.
-     */
-    public static <T> T setListeners(T instance, Collection<MethodCallListener> listeners) {
-        try {
-            LISTENERS_GUARD.acquire();
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
-        try {
-            int i = 0;
-            boolean wasInstancePresent = false;
-            while (i < LISTENERS.size()) {
-                Pair<WeakReference<?>, Collection<MethodCallListener>> pair = LISTENERS.get(i);
-                Object key = pair.getKey().get();
-                if (key == null) {
-                    // The instance has been garbage-collected
-                    LISTENERS.remove(i);
-                    continue;
-                }
-
-                if (key == instance) {
-                    pair.getValue().clear();
-                    pair.getValue().addAll(listeners);
-                    wasInstancePresent = true;
-                }
-                i++;
-            }
-            if (!wasInstancePresent) {
-                LISTENERS.add(new Pair<>(new WeakReference<>(instance), new HashSet<>(listeners)));
-            }
-        } finally {
-            LISTENERS_GUARD.release();
-        }
-        return instance;
-    }
-
-    /**
-     * Fetches listeners for the particular proxied instance.
-     *
-     * @param instance The proxied instance.
-     */
-    public static Collection<MethodCallListener> getListeners(Object instance) {
-        try {
-            LISTENERS_GUARD.acquire();
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
-        try {
-            int i = 0;
-            Collection<MethodCallListener> result = Collections.emptySet();
-            while (i < LISTENERS.size()) {
-                Pair<WeakReference<?>, Collection<MethodCallListener>> pair = LISTENERS.get(i);
-                Object key = pair.getKey().get();
-                if (key == null) {
-                    // The instance has been garbage-collected
-                    LISTENERS.remove(i);
-                    continue;
-                }
-
-                if (key == instance) {
-                    result = pair.getValue();
-                }
-                i++;
-            }
-            return result;
-        } finally {
-            LISTENERS_GUARD.release();
-        }
-    }
 
     /**
      * A magic method used to wrap public method calls in classes
@@ -156,7 +50,7 @@ public class Interceptor {
             @AllArguments Object[] args,
             @SuperCall Callable<?> callable
     ) throws Throwable {
-        Collection<MethodCallListener> listeners = getListeners(self);
+        Collection<MethodCallListener> listeners = ProxyListenersContainer.getInstance().getListeners(self);
         if (listeners.isEmpty()) {
             return callable.call();
         }
