+++ src/test/java/io/appium/java_client/proxy/ProxyHelpersTest.java
@@ -93,7 +93,6 @@ class ProxyHelpersTest {
             }
         };
         RemoteWebDriver driver = createProxy(RemoteWebDriver.class, Collections.singletonList(listener));
-
         assertThrows(
                 IllegalStateException.class,
                 () -> driver.get("http://example.com/")
